pyproject.toml
[project]
name = "padel-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.0",
]
